Rename low-info contract modules
