+++ packages/extension/src/webview/components/graph/debug/api.ts
@@ -1,5 +1,5 @@
 import { useEffect, type MutableRefObject, type RefObject } from 'react';
-import type { GraphDebugControls } from './contracts/types';
+import type { GraphDebugControls } from './contracts/protocol';
 import { installGraphDebugApi } from './install';
 import type { DebugNode } from './snapshot';
