+++ packages/extension/src/webview/components/nodeTooltip/view.tsx
@@ -15,7 +15,7 @@ import {
 import { cn } from '../ui/cn';
 import { Separator } from '../ui/separator';
 import { TooltipHeader, TooltipStats, TooltipExtraSections } from './content';
-import type { TooltipAction } from '../../pluginHost/api/contracts/types';
+import type { TooltipAction } from '../../pluginHost/api/contracts/webview';
 import type { WebviewPluginHost } from '../../pluginHost/manager';
 import { SlotHost } from '../../pluginHost/slotHost/view';
