+++ packages/extension/src/webview/pluginHost/api.ts
@@ -10,8 +10,8 @@ import type {
   TooltipProviderFn,
   WebviewDisposable,
   CodeGraphyWebviewAPI,
-} from './api/contracts/types';
-import type { BadgeOpts, RingOpts, LabelOpts } from './api/contracts/types';
+} from './api/contracts/webview';
+import type { BadgeOpts, RingOpts, LabelOpts } from './api/contracts/webview';
 
 type GraphInteractionMessage = {
   type: 'GRAPH_INTERACTION';
