+++ packages/extension/src/webview/pluginHost/api/registration/index.ts
@@ -3,8 +3,8 @@
  * @module webview/pluginHost/api/registration
  */
 
-import type { NodeRenderFn, OverlayRenderFn, TooltipProviderFn, WebviewDisposable } from '../contracts/types';
-import type { GraphPluginSlot } from '../contracts/types';
+import type { NodeRenderFn, OverlayRenderFn, TooltipProviderFn, WebviewDisposable } from '../contracts/webview';
+import type { GraphPluginSlot } from '../contracts/webview';
 import {
   attachSlotHost as attachSlotHostImpl,
   detachSlotHost as detachSlotHostImpl,
