+++ packages/extension/src/webview/pluginRuntime/moduleResolver.ts
@@ -3,7 +3,7 @@
  * @module webview/pluginModuleResolver
  */
 
-import type { CodeGraphyWebviewAPI } from '../pluginHost/api/contracts/types';
+import type { CodeGraphyWebviewAPI } from '../pluginHost/api/contracts/webview';
 
 export interface PluginWebviewModule {
   activate?: (api: CodeGraphyWebviewAPI) => void | Promise<void>;
