+++ packages/extension/tests/webview/graph/rendering/nodes/canvas3d.test.ts
@@ -21,13 +21,13 @@ vi.mock('../../../../../src/webview/components/graph/rendering/shapes/draw/three
   createNodeMesh: vi.fn(() => ({ kind: 'mesh' })),
 }));
 
-vi.mock('../../../../../src/webview/components/graph/support/contracts/types', () => ({
+vi.mock('../../../../../src/webview/components/graph/support/contracts/forceGraph', () => ({
   setSpriteVisible: vi.fn(),
 }));
 
 import SpriteText from 'three-spritetext';
 import { createImageSprite, createNodeMesh } from '../../../../../src/webview/components/graph/rendering/shapes/draw/threeDimensional';
-import { setSpriteVisible } from '../../../../../src/webview/components/graph/support/contracts/types';
+import { setSpriteVisible } from '../../../../../src/webview/components/graph/support/contracts/forceGraph';
 import type { FGNode } from '../../../../../src/webview/components/graph/model/build';
 import { createNodeThreeObject } from '../../../../../src/webview/components/graph/rendering/nodes/canvas3d';
