+++ packages/extension/tests/webview/pluginHost/api/registration/containers.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import type { GraphPluginSlot } from '../../../../../src/webview/pluginHost/api/contracts/types';
+import type { GraphPluginSlot } from '../../../../../src/webview/pluginHost/api/contracts/webview';
 import {
   attachSlotHost,
   detachSlotHost,
