+++ packages/extension/tests/webview/pluginHost/api/tooltip.mutations.test.ts
@@ -5,7 +5,7 @@
  */
 import { describe, it, expect, vi, afterEach } from 'vitest';
 import { aggregateTooltipContent } from '../../../../src/webview/pluginHost/api/tooltip';
-import type { TooltipProviderFn, TooltipContext } from '../../../../src/webview/pluginHost/api/contracts/types';
+import type { TooltipProviderFn, TooltipContext } from '../../../../src/webview/pluginHost/api/contracts/webview';
 
 const context: TooltipContext = {
   node: { id: 'src/App.ts', label: 'App', color: '#ffffff' },
