+++ packages/extension/tests/webview/pluginHost/api/tooltip.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi } from 'vitest';
 import { aggregateTooltipContent } from '../../../../src/webview/pluginHost/api/tooltip';
-import type { TooltipProviderFn, TooltipContext } from '../../../../src/webview/pluginHost/api/contracts/types';
+import type { TooltipProviderFn, TooltipContext } from '../../../../src/webview/pluginHost/api/contracts/webview';
 
 const context: TooltipContext = {
   node: { id: 'src/App.ts', label: 'App', color: '#ffffff' },
