+++ packages/extension/tests/webview/pluginHost/manager.test.ts
@@ -1,5 +1,5 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import type { TooltipProviderFn } from '@/webview/pluginHost/api/contracts/types';
+import type { TooltipProviderFn } from '@/webview/pluginHost/api/contracts/webview';
 import { WebviewPluginHost } from '../../../src/webview/pluginHost/manager';
 
 function createMockContext(): CanvasRenderingContext2D {
